avox_services/knowledge_service/requirements.txt
# Основные
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1

# Асинхронные HTTP/SQL
aiohttp==3.12.14
httpx==0.28.1
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.41
psycopg[binary]==3.2.9

# Векторные операции и ML
nltk==3.9.1
sentence-transformers==5.0.0
huggingface_hub==0.33.4
pgvector==0.4.1
numpy==2.3.1
torch==2.7.1

# LangChain Core
langchain-core==0.3.68
langchain-community==0.3.27
langchain-text-splitters==0.3.8
langchainhub==0.1.21

# OpenAI / OpenRouter
openai==1.95.1
requests==2.32.4

# Вспомогательные
sqlalchemy-utils==0.41.2
alembic==1.16.4
python-dateutil==2.9.0.post0
pytz==2025.2

# Типы (для mypy)
pydantic==2.11.7
types-pytz==2025.2.0.20250516
types-python-dateutil==2.9.0.20250708
types-requests==2.32.4.20250611

# Для статической проверки (опционально)
mypy==1.5.1  # нужен только если используешь mypy для статической проверки
